+++ src/pages/index.tsx
@@ -1,3 +1,5 @@
+import IconButton from "@/components/common/Button/IconButton";
+
 export default function Landing() {
-  return <div>Landing </div>;
+  return <div><IconButton variant="filled">관리하기</IconButton></div>;
 }
